Split shared/reporting.py into a layered package
reporting.py had grown to 1,929 lines spanning value formatting, CSV
writing, text summaries, summary.json construction, result-bundle
publication, and post-training artifact generation. Split it into
submodules that layer bottom-up, with the package __init__ re-exporting
the same public surface so no caller changes.

  formatting.py       value normalisation + duration formatting (leaf)
  gates.py            curriculum-gate evaluation
  csv_output.py       CSV schema, shared writer, evaluation evidence
  text_summaries.py   stage_summary.txt / training_summary.txt
  summaries.py        schema-v2 summary.json + provenance
  bundles.py          portable result bundle publication
  stage_artifacts.py  SB3 and JAX/MJX post-training artifacts

Also drop _backend_name(), which was defined but never called, and point
test_reporting.py at reporting.csv_output for the private
_compute_fieldnames helper rather than re-exporting it from the package.
