+++ environments/shared/reporting/gates.py
@@ -1,0 +1,49 @@
+"""Curriculum-gate evaluation against recorded evaluation history."""
+
+from __future__ import annotations
+
+from typing import Any
+
+
+def evaluate_recorded_gate(
+    curriculum: dict[str, Any],
+    evaluations: list[dict[str, Any]],
+) -> bool | None:
+    """Evaluate a curriculum gate only when every enabled metric is recorded.
+
+    Evaluations must be chronological. ``None`` means the available records
+    cannot prove either a pass or a failure—for example, when a velocity or
+    success-rate gate is enabled but that metric was not saved.
+    """
+    criteria: list[tuple[str, float]] = []
+    if curriculum.get("min_avg_reward") is not None:
+        criteria.append(("mean_reward", float(curriculum["min_avg_reward"])))
+    if curriculum.get("min_avg_episode_length") is not None:
+        criteria.append(("mean_episode_length", float(curriculum["min_avg_episode_length"])))
+    if float(curriculum.get("min_avg_forward_vel") or 0.0) > 0.0:
+        criteria.append(("mean_forward_vel", float(curriculum["min_avg_forward_vel"])))
+    if float(curriculum.get("min_success_rate") or 0.0) > 0.0:
+        criteria.append(("mean_success_rate", float(curriculum["min_success_rate"])))
+    if not criteria or not evaluations:
+        return None
+
+    min_eval_episodes = int(curriculum.get("min_eval_episodes", 10))
+    required_consecutive = int(curriculum.get("required_consecutive", 3))
+    consecutive = 0
+    incomplete = False
+    for evaluation in evaluations:
+        required_values = [evaluation.get(key) for key, _ in criteria]
+        n_episodes = evaluation.get("n_episodes")
+        if any(value is None for value in required_values) or n_episodes is None:
+            incomplete = True
+            consecutive = 0
+            continue
+
+        passes = int(n_episodes) >= min_eval_episodes and all(
+            float(evaluation[key]) >= threshold for key, threshold in criteria
+        )
+        consecutive = consecutive + 1 if passes else 0
+        if consecutive >= required_consecutive:
+            return True
+
+    return None if incomplete else False
